# myphoenixphone/apps/backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /eligibility:
    get:
      description: Verifies if a device is eligible for the buyback program by
        checking SIM swap history and reachability status. Requires valid user
        consent.
      operationId: EligibilityController_checkEligibility
      parameters:
        - name: phoneNumber
          required: true
          in: query
          description: Phone number in E.164 format (e.g., +33612345678)
          schema:
            example: "+33612345678"
      responses:
        "200":
          description: Eligibility evaluation completed successfully
          content:
            application/json:
              schema:
                example:
                  eligible: true
                  reasons:
                    - SIM_SWAP_RECENT
                    - DEVICE_REACHABLE
                    - MEETS_CRITERIA
                  snapshot:
                    simSwap:
                      swappedAt: 2025-10-16T10:30:00.000Z
                      daysAgo: 5
                    reachability:
                      reachable: true
                      connectivity:
                        - DATA
        "401":
          description: Valid consent required - user has not granted consent
      summary: Check device eligibility
      tags: &a1
        - eligibility
  /eligibility/signals:
    get:
      description: Retrieves stored eligibility signals for debugging and audit
        purposes. Requires valid consent.
      operationId: EligibilityController_getStoredSignals
      parameters:
        - name: phoneNumber
          required: true
          in: query
          description: Phone number in E.164 format
          schema:
            example: "+33612345678"
      responses:
        "200":
          description: Stored signals retrieved successfully
      summary: Get stored eligibility signals
      tags: *a1
...
  /workers/cohorts/{name}/members:
    get:
      operationId: WorkersController_getCohortMembers
      parameters:
        - name: name
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: true
          in: query
          schema:
            type: string
        - name: limit
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a8
  /workers/runs:
    get:
      operationId: WorkersController_getWorkerRuns
      parameters:
        - name: limit
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a8
...
  /workers/daily-refresh:
    post:
      operationId: WorkersController_triggerDailyRefresh
      parameters:
        - name: triggeredBy
          required: true
          in: query
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags: *a8
...
info:
  title: MyPhoenixPhone API
  description: API for MyPhoenixPhone - Dormant device buyback platform using
    Orange Network APIs
  version: "1.0"
  contact: {}
...
servers:
  - url: http://localhost:3003
    description: Development server
  - url: https://api.myphoenixphone.com
    description: Production server
...